+++ src/Command/Ingester/LineItemIngesterCommand.php
@@ -186,6 +186,7 @@ EOF
             $rawLineItem['slug'],
             LineItem::STATUS_ENABLED,
             (int)$rawLineItem['maxAttempts'],
+            $rawLineItem['groupId'] ?? null,
             $startAt,
             $endAt
         );